+++ lib/src/reader/excel_parser.dart
@@ -8,6 +8,7 @@ class Parser extends _ParserBase
         _ParserRelationsMixin,
         _ParserDrawingsMixin,
         _ParserCommentsMixin,
+        _ParserTablesMixin,
         _ParserWorksheetFeaturesMixin {
   Parser._(super.excel);
 
@@ -97,6 +98,7 @@ class Parser extends _ParserBase
     _parseHyperlinksForSheet(sheetName);
     _parseDrawingsForSheet(sheetName);
     _parseCommentsForSheet(sheetName);
+    _parseTablesForSheet(sheetName);
     _parseDataValidationsForSheet(sheetName);
     _parseSheetViewForSheet(sheetName);
     _parseAutoFilterForSheet(sheetName);